refactor(ui): extract settings/data/value-helpers.ts — pure config-reading helpers
